Apply `ethlike` types to logging wrapper
Removed all references to `go-ethereum` and
used `ethlike` types instead. This change should
make the logging wrapper ready to work will all
ETH-like chains.
